requirements.txt
absl-py==0.7.1
astor==0.7.1
bleach==1.5.0
certifi==2018.8.24
Click==7.0
cycler==0.10.0
Cython==0.29.6
dlib==19.17.0
enum34==1.1.6
face-recognition==1.2.3
face-recognition-models==0.3.0
gast==0.2.2
grpcio==1.19.0
html5lib==0.9999999
imutils==0.5.2
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==3.0.0
mkl-fft==1.0.0
mkl-random==1.0.1
numpy==1.15.2
olefile==0.45.1
Pillow==5.2.0
protobuf==3.5.2.post1
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
tornado==5.1.1
tqdm==4.31.1
Werkzeug==0.14.1
